Reject deletion of in-use L2Policy or L3Policy
Before deleting an L2Policy, check that no PolicyTargetGroups
reference it. Before deleting an L3Policy, ensure that no L2Policies
reference it. Raise exceptions if they are in-use.

Closes-bug: 1415244
Closes-bug: 1407323

+++ gbpservice/neutron/db/grouppolicy/group_policy_db.py
@@ -1179,6 +1179,8 @@ class GroupPolicyDbPlugin(gpolicy.GroupPolicyPluginBase,
     def delete_l2_policy(self, context, l2_policy_id):
         with context.session.begin(subtransactions=True):
             l2p_db = self._get_l2_policy(context, l2_policy_id)
+            if l2p_db.policy_target_groups:
+                raise gpolicy.L2PolicyInUse(l2_policy_id=l2_policy_id)
             context.session.delete(l2p_db)
 
     @log.log
